Move charcoal/email/email-log to charcoal/email/objects/email-log. Keep old class for compatibility.
