+++ metadata/charcoal/email/objects/email-log.json
@@ -1,7 +1,8 @@
 {
     "properties": {
         "id": {
-            "type": "id"
+            "type": "id",
+            "mode": "uuid"
         },
         "queue_id": {
             "type": "string",
